style(web): consolidate color semantics
The app mixed three oranges (--primary, --accent, --ring) for the same
'primary action' role and hardcoded chart/warning colors that drifted from
the theme tokens.

- button.tsx: the default variant now uses bg-accent (the vivid orange already
  used by every primary CTA) instead of bg-primary, and its shadow tones were
  updated to match. Per-route bg-accent overrides (transactions, merchants,
  categories, landing CTAs, split dialog) are removed.
- top-nav.tsx: active nav state switched from primary to accent so it matches
  the mobile drawer's active state.
- index.css: added --warning/--warning-muted tokens (light + dark) and
  registered them with @theme inline.
- badge.tsx: the warning variant now uses the warning tokens instead of
  hardcoded amber-* classes.
- stats.tsx: savings-rate warning color uses text-warning.
- lib/chart-colors.ts: shared categorical palette + hashing, replacing the
  duplicated arrays in category-pie-chart.tsx and income-expense-sankey.tsx.
- income-expense-sankey.tsx: income/saved/expense link colors now reference
  the --income/--savings/--expense theme tokens instead of hardcoded hex.
