+++ apps/web/src/index.css
@@ -216,6 +216,8 @@ body {
   --expense-muted: #fee2e2;
   --savings: #0d9488;
   --savings-muted: #ccfbf1;
+  --warning: #d97706;
+  --warning-muted: #fef3c7;
 
   /* ==========================================
      CHART COLORS - Vibrant Palette
@@ -305,6 +307,8 @@ body {
   --expense-muted: #450a0a;
   --savings: #2dd4bf;
   --savings-muted: #0f766e;
+  --warning: #fbbf24;
+  --warning-muted: #78350f;
 
   /* ==========================================
      CHART COLORS - Dark Mode (vibrant)
@@ -398,6 +402,8 @@ body {
   --color-income-muted: var(--income-muted);
   --color-expense-muted: var(--expense-muted);
   --color-savings-muted: var(--savings-muted);
+  --color-warning: var(--warning);
+  --color-warning-muted: var(--warning-muted);
 }
 
 @layer base {
