+++ apps/web/src/routes/categories.tsx
@@ -116,7 +116,7 @@ function RouteComponent() {
         actions={
           <Dialog open={open} onOpenChange={setOpen}>
             <DialogTrigger asChild>
-              <Button className="bg-accent hover:bg-accent/90 text-accent-foreground font-medium">
+              <Button>
                 <Plus className="w-4 h-4 mr-2" />
                 New category
               </Button>
