+++ apps/web/src/routes/merchants.tsx
@@ -141,7 +141,7 @@ function RouteComponent() {
             </AlertDialog>
             <Dialog open={open} onOpenChange={setOpen}>
               <DialogTrigger asChild>
-                <Button className="bg-accent hover:bg-accent/90 text-accent-foreground font-medium">
+                <Button>
                   <Plus className="w-4 h-4 mr-2" />
                   New merchant
                 </Button>
